moves services into flat folder structure
