+++ client/src/app/modules/account/auth/register/register.component.ts
@@ -3,8 +3,8 @@ import { FormBuilder, FormGroup, Validators, AbstractControl } from '@angular/fo
 import { Observable, combineLatest } from 'rxjs';
 
 import { CustomValidators } from '@utility/custom-validators';
-import { AuthService } from '@services/auth/auth.service';
-import { ErrorService } from '@services/error/error.service';
+import { AuthService } from '@services/auth.service';
+import { ErrorService } from '@services/error.service';
 import { EApi } from '@models/enums/api.enum';
 import { map } from 'rxjs/operators';
