+++ client/src/app/modules/account/profile/profile.component.ts
@@ -3,7 +3,7 @@ import { Router } from '@angular/router';
 import { Observable } from 'rxjs';
 
 import { Auth } from '@models/classes/auth';
-import { AuthService } from '@services/auth/auth.service';
+import { AuthService } from '@services/auth.service';
 
 @Component({
   selector: 'app-profile',
