+++ client/src/app/modules/learn/quiz/quiz-cards/quiz-cards.component.ts
@@ -4,7 +4,7 @@ import { map, first } from 'rxjs/operators';
 
 import { ICountry } from '@models/interfaces/country.interface';
 import { EQuizType } from '@models/enums/quiz-type.enum';
-import { QuizService } from '@services/quiz/quiz.service';
+import { QuizService } from '@services/quiz.service';
 import { staggerAnimation, fadeInAnimation } from '@utility/animations';
 import { shuffle } from "lodash-es";
