+++ client/src/app/modules/learn/quiz/quiz-menu/quiz-menu.component.ts
@@ -8,7 +8,7 @@ import { Quiz } from '@models/classes/quiz';
 import { EQuizType } from '@models/enums/quiz-type.enum';
 import { ERoute } from '@models/enums/route.enum';
 import { FixedSlideablePanelPosition } from '@shared/components/fixed-slideable-panel/fixed-slideable-panel.component';
-import { QuizService } from '@services/quiz/quiz.service';
+import { QuizService } from '@services/quiz.service';
 import { ICountry } from '@models/interfaces/country.interface';
 
 interface IViewModel {
